Move user list and book list to separate components
